moved file testing into io file
